RoboAssistant_Transfer/requirements.txt
# RoboAssistant Dependencies

# Core libraries
numpy>=1.19.0
matplotlib>=3.3.0
opencv-python>=4.5.0
pyserial>=3.5
pillow>=8.0.0

# Voice recognition and speech synthesis
SpeechRecognition>=3.8.1
pyttsx3>=2.90
nltk>=3.6.2

# Intel RealSense SDK
pyrealsense2>=2.50.0

# RPLIDAR SDK
rplidar>=0.9.2

# TensorFlow and Coral libraries
# Note: Install the appropriate wheel for your platform
# See https://coral.ai/docs/accelerator/get-started/ for Coral setup
tflite-runtime>=2.5.0
pycoral>=2.0.0

# 3D Processing and QC tools
open3d>=0.15.1
trimesh>=3.9.0
pyglet>=1.5.0  # For trimesh visualization

# General utilities
tqdm>=4.62.0
psutil>=5.8.0
requests>=2.26.0
PyYAML>=5.4.1
jsonschema>=3.2.0

# Optional: GUI
tkinter  # Usually comes with Python
PyQt5>=5.15.0  # Alternative GUI framework

# Optional: data processing
pandas>=1.3.0
scipy>=1.7.0 